% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BloquesAum.R
\name{bloquesaum}
\alias{bloquesaum}
\title{Dise침o de bloques aumentados}
\usage{
bloquesaum(data, block, std_trt, new_trt, response)
}
\arguments{
\item{data}{un conjunto de datos}

\item{block}{un factor de bloque}

\item{std_trt}{un tratamiento est치ndar}

\item{new_trt}{un tratamiento nuevo}

\item{response}{una variable de respuesta}
}
\value{
un data frame con lo valores de la suma de cuadrados, los grados de libertad, el estadistico F y el p. valor
}
\description{
Dise침o de bloques aumentados
}
\examples{
\dontrun{
# Ejemplo de uso de la funci칩n con los datos del ejemplo
data <- data.frame(block = rep(1:3, each = 4),
           std_trt = rep(c("A", "B"), times = 6),
           new_trt = rep(c("C", "D"), each = 6),
           response = c(1.23, 2.34, 3.45, 4.56, 5.67, 6.78, 7.89, 8.90, 9.01, 0.12, 1.23, 2.34))
bloquesaum(data, "block", "std_trt", "new_trt", "response")
}
}
